osa 4 done

+++ osa4/blog/controllers/blogs.js
@@ -2,6 +2,7 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const { response } = require('express')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', {username: 1, name: 1})
@@ -19,10 +20,25 @@ blogsRouter.get('/:id', async(request,response) => {
 
 })
 
-blogsRouter.post('/', async (request, response) => {
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
+
+
+blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
 
-    const user = await User.findById(body.userId)
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
    if (body.title === undefined || body.url === undefined){
        return response.status(400).json({error: 'title or url missing'}).end()
@@ -45,8 +61,20 @@ blogsRouter.post('/', async (request, response) => {
   })
 
   blogsRouter.delete('/:id', async(request, response, next) => {
-
       const id = request.params.id
+      const token = getTokenFrom(request)
+      const decodedToken = jwt.verify(token, process.env.SECRET)
+      
+      const blog = await Blog.findById(id)
+      const blogUser = blog.user.toString()
+      console.log(blogUser)
+      const tokenUser = decodedToken.id
+      console.log(tokenUser)
+
+      if(!token || !decodedToken.id || blogUser !== tokenUser) {
+        return response.status(401).json({ error: 'Invalid token or user'})
+      }
+    
 
       await Blog.findByIdAndRemove(id)
       response.status(204).end()
